Outras exceções

+++ src/main/java/com/dvsuperior/cliente/services/ClientService.java
@@ -3,11 +3,14 @@ package com.dvsuperior.cliente.services;
 import com.dvsuperior.cliente.dto.ClientDTO;
 import com.dvsuperior.cliente.entities.Client;
 import com.dvsuperior.cliente.repositories.ClientRepository;
+import com.dvsuperior.cliente.services.exceptions.DatabaseException;
 import com.dvsuperior.cliente.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -40,15 +43,28 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }
+        catch (ResourceNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
@@ -59,9 +75,6 @@ public class ClientService {
         entity.setChildren(dto.getChildren());
     }
 }
-
-
-
